+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 * [CHANGE] Renamed `sourceInfo` label to `info`
 * [CHANGE] `summary` is read from a label (not annotation anymore) in rules file
 * [CHANGE] `description` is read from a label if exists, otherwise from annotation
-* [FEATURE] Target view
+* [FEATURE] Target view (#10)
+* [FEATURE] Environment variable substitution (#11)
 * [ENHANCEMENT] Added security constraints for restricted access to some directories (web.xml)
 * [ENHANCEMENT] URLs without .xhtml extension
 * [ENHANCEMENT] Supported new labels `team`, `currentValue`, `url`, `eventType`, `probableCause`